+++ server/component/style/radio/tpl_radio.php
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 ?>
 <div class="form-check <?php echo $inline; ?>">
-    <input class="form-check-input" type="radio" name="<?php echo $this->name; ?>" value="<?php echo $value; ?>" <?php echo $checked; ?>>
+    <input class="selfhelpRadio form-check-input" data-locked_after_submit="<?php echo $this->locked_after_submit; ?>" type="radio" name="<?php echo $this->name; ?>" value="<?php echo $value; ?>" <?php echo $required; ?> <?php echo $checked; ?>>
     <label class="form-check-label">
         <?php echo $text; ?>
     </label>