+++ server/component/style/textarea/js/textarea.js
@@ -1,3 +1,12 @@
 $(document).ready(function() {
     autosize($('textarea'));
+    check_textarea_locked_after_submit();
 });
+
+function check_textarea_locked_after_submit(){
+    $('.selfhelpTextArea').each(function(){
+        if($(this).data('locked_after_submit') && $(this).val()){
+            $(this).prop('readonly', true);
+        }        
+    })
+}